start refactoring and renaming to clear up code structure
