+++ dominion_dividers.py
@@ -1,0 +1,6 @@
+import domdiv
+import os
+import sys
+
+if __name__ == '__main__':
+    domdiv.main(sys.argv[1:], os.path.dirname(__file__))
